+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
-# idea-first-plugin Changelog
+# adl-idea Changelog
 
 ## [Unreleased]
+### Added
+- Multiple servers
+- raw request / response log
+- autocomplete using server metadata
+- multiple aqls per .aql file
+
+## 0.0.1
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)